Add ActivationQuantizationHolder and use it during GPTQ
Add ActivationQuantizationHolder layer to hold an activation quantizer and quantize during inference.
Update Keras model builder to use it as quantization layers if it is passed (instead of
KerasQuantizeWrapper that was used for both weights and activations).
For now, only GPTQ uses it by passing it to the model builder - thus, GPTQ models wrap layers
with weight quantizers and quantize activations using ActivationQuantizationHolder layers.